# src/main/resources/static/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI specification - Austin Keiber",
        "description": "OpenAPI Documentation for Blogging API",
        "contact": {
            "name": "Austin Keiber",
            "email": "adkeiber@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local ENV"
        }
    ],
    "paths": {
        "/tag/update/{id}": {
            "post": {
                "tags": [
                    "Blog Tag Controller"
                ],
                "operationId": "updateById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Tag"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tag/create": {
            "post": {
                "tags": [
                    "Blog Tag Controller"
                ],
                "summary": "Endpoint that takes in and saves a tag object.\r\nIf a tag already exists with the same value or the tag value is empty/null the tag will not be saved.",
                "description": "Endpoint to create a new Tag",
                "operationId": "create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Tag"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post/update/{id}": {
            "post": {
                "tags": [
                    "Blog Post Controller"
                ],
                "operationId": "updateById_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post/create": {
            "post": {
                "tags": [
                    "Blog Post Controller"
                ],
                "operationId": "create_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tag/getByValue/{value}": {
            "get": {
                "tags": [
                    "Blog Tag Controller"
                ],
                "operationId": "getByValue",
                "parameters": [
                    {
                        "name": "value",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tag/getById/{id}": {
            "get": {
                "tags": [
                    "Blog Tag Controller"
                ],
                "operationId": "getById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tag/getAll": {
            "get": {
                "tags": [
                    "Blog Tag Controller"
                ],
                "operationId": "getAllTags",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post/getByTitle/{title}": {
            "get": {
                "tags": [
                    "Blog Post Controller"
                ],
                "operationId": "getByTitle",
                "parameters": [
                    {
                        "name": "title",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post/getByTagValue/{tag}": {
            "get": {
                "tags": [
                    "Blog Post Controller"
                ],
                "operationId": "getByTagValue",
                "parameters": [
                    {
                        "name": "tag",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post/getById/{id}": {
            "get": {
                "tags": [
                    "Blog Post Controller"
                ],
                "operationId": "getById_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post/getByDate/{date}": {
            "get": {
                "tags": [
                    "Blog Post Controller"
                ],
                "operationId": "getByDate",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post/getAll": {
            "get": {
                "tags": [
                    "Blog Post Controller"
                ],
                "operationId": "getAll",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tag/deleteByValue/{value}": {
            "delete": {
                "tags": [
                    "Blog Tag Controller"
                ],
                "operationId": "deleteByValue",
                "parameters": [
                    {
                        "name": "value",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tag/deleteById/{id}": {
            "delete": {
                "tags": [
                    "Blog Tag Controller"
                ],
                "operationId": "deleteById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post/deleteById/{id}": {
            "delete": {
                "tags": [
                    "Blog Post Controller"
                ],
                "operationId": "deleteById_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Tag": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string"
                    }
                }
            },
            "Post": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "contents": {
                        "type": "string"
                    },
                    "publishDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tag"
                        }
                    }
                }
            },
            "Api Error": {
            	"type": "object",
            	"properties": {
            		"status": {
                    	"type": "HttpStatus"
                	},
                	"timeStamp": {
                    	"type": "LocalDateTime"
                	},
            		"message": {
                        "type": "string"
                    },
                	"debugMessage": {
                    	"type": "string"
                	}
            	}
            }
        }
    }
}